Follow up UAT
- Apply carousel only to Actions List block
- Improve styles
- Reduce the complexity of toolbar
