+++ assets/src/blocks/editorIndex.js
@@ -3,7 +3,7 @@ import {registerActionsList} from './ActionsList';
 import {registerSubmenuBlock} from './Submenu/SubmenuBlock';
 import {registerTakeActionBoxoutBlock} from './TakeActionBoxout/TakeActionBoxoutBlock';
 import {registerHappypointBlock} from './Happypoint/HappypointBlock';
-import {setupQueryBlockExtension} from '../block-editor/QueryBlockExtension';
+import {setupQueryBlockExtension} from '../block-editor/QueryBlockExtension/QueryBlockExtension';
 
 wp.domReady(() => {
   // Make sure to unregister the posts-list native variation before registering planet4-blocks/posts-list-block
