+++ assets/src/scss/layout/_query-loop-overrides.scss
@@ -18,13 +18,15 @@
     }
 
     .wp-block-post {
-      box-shadow: 0 1px 8px rgba($grey-900, 0.16);
+      display: flex;
+      flex-direction: column;
+      margin-bottom: 0;
+      min-height: 490px;
       transition: box-shadow 0.2s;
       border-radius: 4px;
       background: var(--white);
+      box-shadow: 0 1px 8px rgba($grey-900, 0.16);
       cursor: pointer;
-      margin-bottom: 0;
-      min-height: 490px;
 
       // Applied to content
       .wp-block-group:first-of-type {
@@ -33,7 +35,13 @@
 
       .wp-block-post-featured-image {
         overflow: hidden;
-        border-radius: 4px 4px 0;
+        border-radius: 4px 4px 0 0;
+      }
+
+      .read-more-nav {
+        display: flex;
+        justify-content: flex-end;
+        margin-top: auto;
       }
 
       &:hover {
